requirements.txt
pygame==2.1.2
PyOpenGL==3.1.5
numpy
PyOpenGL_accelerate